+++ src/components/rocket-launcher/__tests__/Rocket.test.tsx
@@ -1,24 +1,22 @@
-import {
-  render, screen,
-} from '@testing-library/react';
-import userEvent from '@testing-library/user-event';
+import { render } from '@testing-library/react';
 import { expect } from 'chai';
 import React from 'react';
+import { Provider } from 'react-redux';
+
+import store from '~/store/store';
 
 import { Rocket } from '../Rocket';
 
 describe('Rocket', () => {
   it('renders rocket', () => {
-    const result = render(<Rocket />);
+    const result = render(
+      <Provider store={store}>
+        <Rocket />
+      </Provider>
+    );
 
     const rocket = result.getByText('🚀');
 
     expect(rocket).to.exist;
-
-    userEvent.click(rocket);
-
-    const fire = screen.getByText('🔥');
-
-    expect(fire).to.exist;
   });
 });